Refactoring Costmap2DPublisher and Costmap2DROS

+++ nav2_costmap_2d/include/nav2_costmap_2d/costmap_2d_publisher.hpp
@@ -72,7 +72,7 @@ public:
     std::string global_frame,
     std::string topic_name,
     bool always_send_full_costmap = false,
-    double map_vis_y = 0.0);
+    float map_vis_y = 0.0);
 
   /**
    * @brief  Destructor
@@ -157,7 +157,7 @@ private:
   double saved_origin_y_;
   bool active_;
   bool always_send_full_costmap_;
-  double map_vis_y_;
+  float map_vis_y_;
 
   // Publisher for translated costmap values as msg::OccupancyGrid used in visualization
   rclcpp_lifecycle::LifecyclePublisher<nav_msgs::msg::OccupancyGrid>::SharedPtr costmap_pub_;
